# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CampusOps Expo API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /students:
    get:
      operationId: getStudents
      tags: [students]
      summary: Get all students (Admin)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of students
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StudentProfile"
    post:
      operationId: createStudent
      tags: [students]
      summary: Create student (Admin)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStudentRequest"
      responses:
        "201":
          description: Student created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
  /students/{id}:
    get:
      operationId: getStudentById
      tags: [students]
      summary: Get student by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Student profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StudentProfile"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    StudentProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        rollNumber:
          type: string
        hostelId:
          type: string
        hostelName:
          type: string
        roomNumber:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - email
        - rollNumber
    CreateStudentRequest:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
        rollNumber:
          type: string
        hostelId:
          type: string
        roomNumber:
          type: string
      required:
        - name
        - email
        - password
        - rollNumber
...